cloud-run/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
umap-learn==0.5.7
hdbscan==0.8.40
scikit-learn==1.5.2
numpy==1.26.4
requests==2.32.3
firebase-admin==6.5.0
python-multipart==0.0.9
